# stackhawk-openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "PrintFlow API",
  "description": "Store-scoped print order queue: CSV intake, PSD composition, TIFF/PDF output, CUPS dispatch.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/stores": {
   "get": {
    "tags": [
     "stores"
    ],
    "summary": "List Stores",
    "description": "Admins see every store. Operators see only their own (needed for headers/filters).",
    "operationId": "list_stores_stores_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "include_inactive",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Include Inactive"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/StoreWithStats"
         },
         "title": "Response List Stores Stores Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "stores"
    ],
    "summary": "Create Store",
    "operationId": "create_store_stores_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StoreCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StoreOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "parameters": []
   }
  },
  "/stores/{store_id}": {
   "patch": {
    "tags": [
     "stores"
    ],
    "summary": "Update Store",
    "operationId": "update_store_stores__store_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "store_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Store Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StoreUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StoreOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "stores"
    ],
    "summary": "Delete Store",
    "operationId": "delete_store_stores__store_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "store_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Store Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/printers": {
   "get": {
    "tags": [
     "printing"
    ],
    "summary": "List Printers",
    "operationId": "list_printers_printers_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/PrinterOut"
         },
         "type": "array",
         "title": "Response List Printers Printers Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": []
   }
  },
  "/orders/{order_id}/print": {
   "post": {
    "tags": [
     "printing"
    ],
    "summary": "Print Order",
    "description": "Render the PSD + text, then deliver it according to `delivery`.\n\nPRINTER dispatches to CUPS from the server. DOWNLOAD marks the order printed\nand leaves the operator to print the file locally \u2014 the only route that works\nwhen the app runs in the cloud and the printer is attached to the operator's\nown machine. PROOF renders the artifacts and leaves the order status alone.\n\nBoth PRINTER and DOWNLOAD flip the order to PRINTED and, if it was already\nPRINTED, count as a reprint.",
    "operationId": "print_order_orders__order_id__print_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "order_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Order Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PrintRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PrintResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/orders/{order_id}/artifact/{kind}": {
   "get": {
    "tags": [
     "printing"
    ],
    "summary": "Get Artifact",
    "operationId": "get_artifact_orders__order_id__artifact__kind__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "order_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Order Id"
      }
     },
     {
      "name": "kind",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Kind"
      }
     },
     {
      "name": "job_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Job Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "Delivery": {
    "type": "string",
    "enum": [
     "PRINTER",
     "DOWNLOAD",
     "PROOF"
    ],
    "title": "Delivery",
    "description": "How the rendered file reaches paper."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "JobKind": {
    "type": "string",
    "enum": [
     "TIFF",
     "PDF"
    ],
    "title": "JobKind"
   },
   "JobStatus": {
    "type": "string",
    "enum": [
     "QUEUED",
     "RENDERED",
     "SENT_TO_PRINTER",
     "DOWNLOADED",
     "FAILED"
    ],
    "title": "JobStatus"
   },
...
   "ModerationStatus": {
    "type": "string",
    "enum": [
     "UNCHECKED",
     "CLEAR",
     "FLAGGED",
     "NEEDS_REVIEW",
     "APPROVED",
     "REJECTED"
    ],
    "title": "ModerationStatus",
    "description": "Where an order's text stands with the LLM brand-safety gate.\n\nStored as a plain varchar rather than a Postgres enum so that adding a\nmember is an additive migration, not an ALTER TYPE."
   },
   "OrderOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "order_ref": {
      "type": "string",
      "title": "Order Ref"
     },
     "store_id": {
      "type": "integer",
      "title": "Store Id"
     },
     "store_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Store Name"
     },
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "sku_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sku Code"
     },
     "brand": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Brand"
     },
     "amount": {
      "type": "string",
      "title": "Amount"
     },
     "print_format_id": {
      "type": "integer",
      "title": "Print Format Id"
     },
     "print_text": {
      "type": "string",
      "title": "Print Text"
     },
     "status": {
      "$ref": "#/components/schemas/OrderStatus"
     },
     "reprint_count": {
      "type": "integer",
      "title": "Reprint Count"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "printed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Printed At"
     },
     "last_error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Error"
     },
     "moderation_status": {
      "$ref": "#/components/schemas/ModerationStatus",
      "default": "UNCHECKED"
     },
     "moderation_categories": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Moderation Categories"
     },
     "moderation_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Moderation Reason"
     },
     "moderation_note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Moderation Note"
     },
     "moderated_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Moderated At"
     },
     "store": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/StoreOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "print_format": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/PrintFormatOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "printed_by": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/UserOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "reviewed_by": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/UserOut"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "id",
     "order_ref",
     "store_id",
     "store_name",
     "city",
     "sku_code",
     "brand",
     "amount",
     "print_format_id",
     "print_text",
     "status",
     "reprint_count",
     "created_at",
     "printed_at",
     "last_error"
    ],
    "title": "OrderOut"
   },
...
   "OrderStatus": {
    "type": "string",
    "enum": [
     "PENDING",
     "PRINTING",
     "PRINTED",
     "FAILED"
    ],
    "title": "OrderStatus"
   },
...
   "PrintJobOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "order_id": {
      "type": "integer",
      "title": "Order Id"
     },
     "kind": {
      "$ref": "#/components/schemas/JobKind"
     },
     "is_reprint": {
      "type": "boolean",
      "title": "Is Reprint"
     },
     "without_text": {
      "type": "boolean",
      "title": "Without Text"
     },
     "printer_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Printer Name"
     },
     "status": {
      "$ref": "#/components/schemas/JobStatus"
     },
     "passes": {
      "type": "integer",
      "title": "Passes"
     },
     "font_size_used": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Font Size Used"
     },
     "cups_job_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cups Job Id"
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "user": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/UserOut"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "id",
     "order_id",
     "kind",
     "is_reprint",
     "without_text",
     "printer_name",
     "status",
     "passes",
     "font_size_used",
     "cups_job_id",
     "error",
     "created_at"
    ],
    "title": "PrintJobOut"
   },
   "PrintRequest": {
    "properties": {
     "kind": {
      "$ref": "#/components/schemas/JobKind",
      "default": "TIFF"
     },
     "printer_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Printer Name"
     },
     "copies": {
      "type": "integer",
      "maximum": 20.0,
      "minimum": 1.0,
      "title": "Copies",
      "default": 1
     },
     "delivery": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Delivery"
       },
       {
        "type": "null"
       }
      ]
     },
     "send_to_printer": {
      "type": "boolean",
      "title": "Send To Printer",
      "default": true
     },
     "without_text": {
      "type": "boolean",
      "title": "Without Text",
      "default": false
     }
    },
    "type": "object",
    "title": "PrintRequest"
   },
   "PrintResponse": {
    "properties": {
     "job": {
      "$ref": "#/components/schemas/PrintJobOut"
     },
     "order": {
      "$ref": "#/components/schemas/OrderOut"
     },
     "tiff_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Tiff Url"
     },
     "pdf_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pdf Url"
     },
     "preview_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Preview Url"
     },
     "warning": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Warning"
     }
    },
    "type": "object",
    "required": [
     "job",
     "order"
    ],
    "title": "PrintResponse"
   },
   "PrinterOut": {
    "properties": {
     "name": {
      "type": "string",
      "title": "Name"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "is_default": {
      "type": "boolean",
      "title": "Is Default",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "name",
     "status"
    ],
    "title": "PrinterOut"
   },
...
   "StoreCreate": {
    "properties": {
     "code": {
      "type": "string",
      "maxLength": 64,
      "minLength": 1,
      "title": "Code"
     },
     "name": {
      "type": "string",
      "maxLength": 255,
      "minLength": 1,
      "title": "Name"
     },
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "code",
     "name"
    ],
    "title": "StoreCreate"
   },
   "StoreOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "code": {
      "type": "string",
      "title": "Code"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "code",
     "name",
     "city",
     "is_active",
     "created_at"
    ],
    "title": "StoreOut"
   },
   "StoreUpdate": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "is_active": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Active"
     }
    },
    "type": "object",
    "title": "StoreUpdate"
   },
   "StoreWithStats": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "code": {
      "type": "string",
      "title": "Code"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "user_count": {
      "type": "integer",
      "title": "User Count",
      "default": 0
     },
     "pending_orders": {
      "type": "integer",
      "title": "Pending Orders",
      "default": 0
     }
    },
    "type": "object",
    "required": [
     "id",
     "code",
     "name",
     "city",
     "is_active",
     "created_at"
    ],
    "title": "StoreWithStats"
   },
...
   "UserOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "full_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Full Name"
     },
     "role": {
      "$ref": "#/components/schemas/Role"
     },
     "store_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Store Id"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "last_login_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Login At"
     },
     "store": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/StoreOut"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "full_name",
     "role",
     "store_id",
     "is_active",
     "created_at"
    ],
    "title": "UserOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 },
 "servers": [
  {
   "url": "http://localhost:8000"
  }
 ]
}